+++ src/main/java/org/synyx/urlaubsverwaltung/security/ldap/InvalidSecurityConfigurationException.java
@@ -1,0 +1,12 @@
+package org.synyx.urlaubsverwaltung.security.ldap;
+
+/**
+ * Is thrown in case of the security configuration (LDAP/AD) seems to be invalid.
+ */
+public class InvalidSecurityConfigurationException extends IllegalStateException {
+
+    InvalidSecurityConfigurationException(String message) {
+
+        super(message);
+    }
+}