+++ src/main/java/org/synyx/urlaubsverwaltung/security/ldap/UnsupportedMemberAffiliationException.java
@@ -1,0 +1,12 @@
+package org.synyx.urlaubsverwaltung.security.ldap;
+
+/**
+ * Is thrown in case of missing member affiliation.
+ */
+public class UnsupportedMemberAffiliationException extends Exception {
+
+    UnsupportedMemberAffiliationException(String message) {
+
+        super(message);
+    }
+}