+++ src/test/java/org/synyx/urlaubsverwaltung/security/ldap/LdapUserServiceImplTest.java
@@ -1,0 +1,42 @@
+package org.synyx.urlaubsverwaltung.security.ldap;
+
+import org.junit.Before;
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.mockito.Mock;
+import org.mockito.junit.MockitoJUnitRunner;
+import org.springframework.ldap.core.LdapTemplate;
+import org.synyx.urlaubsverwaltung.security.SecurityConfigurationProperties;
+import org.synyx.urlaubsverwaltung.security.ldap.LdapUserMapper;
+import org.synyx.urlaubsverwaltung.security.ldap.LdapUserService;
+import org.synyx.urlaubsverwaltung.security.ldap.LdapUserServiceImpl;
+
+import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.ArgumentMatchers.eq;
+import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.verify;
+
+@RunWith(MockitoJUnitRunner.class)
+public class LdapUserServiceImplTest {
+
+    private LdapUserService sut;
+
+    @Mock
+    private LdapTemplate ldapTemplate;
+    @Mock
+    private LdapUserMapper ldapUserMapper;
+
+    @Before
+    public void setUp() {
+
+        sut = new LdapUserServiceImpl(ldapTemplate, ldapUserMapper, new SecurityConfigurationProperties());
+    }
+
+    @Test
+    public void ensureUsesLdapTemplateToFetchUsers() {
+
+        sut.getLdapUsers();
+
+        verify(ldapTemplate).search(any(), eq(ldapUserMapper));
+    }
+}