+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Migrate database to mariadb [#518](https://github.com/synyx/urlaubsverwaltung/pull/518)
 * Add configuration properties for security and some cleanups [#696](https://github.com/synyx/urlaubsverwaltung/pull/696)
 * Upgrade spring boot to 2.1.5 [#710](https://github.com/synyx/urlaubsverwaltung/pull/710)
+* Migrate double used configuration paths for uv.security [#712](https://github.com/synyx/urlaubsverwaltung/pull/712)
 
 ### master/unreleased
 * Quickfix for better overview about department members: add link to department